refactor: split Source API response types into src/source_api.rs
registry.rs held both SourceCoopRegistry (the BucketRegistry impl) and the
Source API DTOs (SourceProduct, DataConnection, Visibility, ...). The types
aren't discoverable under "registry"; move them to a dedicated source_api
module. registry resolves them, cache fetches them — both now import from
crate::source_api. Pure move; no behavior change.
